More work on SQL grammar AST (TSQL table constraints)

+++ bsn.ModuleStore/Sql/Script/IntegerLiteral.cs
@@ -5,10 +5,10 @@ using bsn.GoldParser.Semantic;
 
 namespace bsn.ModuleStore.Sql.Script {
 	[Terminal("IntegerLiteral")]
-	public class IntegerLiteral: NumericLiteral<int> {
-		public IntegerLiteral(string value): this(int.Parse(value, NumberFormatInfo.InvariantInfo)) {}
+	public class IntegerLiteral: NumericLiteral<long> {
+		public IntegerLiteral(string value): this(long.Parse(value, NumberFormatInfo.InvariantInfo)) {}
 
-		protected IntegerLiteral(int value): base(value) {}
+		protected IntegerLiteral(long value): base(value) {}
 
 		public override double AsDouble {
 			get {
